prompt for apps

+++ ui/components/MesheryApplications.js
@@ -18,6 +18,7 @@ import UploadIcon from "@material-ui/icons/Publish";
 import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 import MUIDataTable from "mui-datatables";
+import PromptComponent from "./PromptComponent";
 import Moment from "react-moment";
 import { withSnackbar } from "notistack";
 import CloseIcon from "@material-ui/icons/Close";
@@ -101,6 +102,7 @@ function MesheryApplications({
   const [search] = useState("");
   const [sortOrder] = useState("");
   const [count, setCount] = useState(0);
+  const modalRef = useRef(null);
   const [pageSize, setPageSize] = useState(10);
   const [applications, setApplications] = useState([]);
   const [selectedRowData, setSelectedRowData] = useState(null);
@@ -371,6 +373,15 @@ function MesheryApplications({
     }
   });
 
+  async function showModal() {
+    let response = await modalRef.current.show({ title : "Delete Aplication?",
+
+      subtitle : "Are you sure you want to delete this application?",
+
+      options : ["YES", "NO"], })
+    return response;
+  }
+
   async function deleteApplication(id) {
     dataFetch(
       `/api/application/${id}`,
@@ -415,9 +426,15 @@ function MesheryApplications({
     download : false,
     customToolbar : CustomToolbar(uploadHandler),
 
-    onRowsDelete : function handleDelete(row) {
-      const fid = Object.keys(row.lookup).map(idx => applications[idx]?.id)
-      fid.forEach(fid => deleteApplication(fid))
+    onRowsDelete : async function handleDelete(row) {
+      let response = await showModal()
+      console.log(response)
+      if (response === "YES") {
+        const fid = Object.keys(row.lookup).map(idx => applications[idx]?.id)
+        fid.forEach(fid => deleteApplication(fid))
+      }
+      if (response === "NO")
+        fetchApplications(page, pageSize, search, sortOrder);
     },
 
     onTableChange : (action, tableState) => {
@@ -476,6 +493,7 @@ function MesheryApplications({
         // @ts-ignore
         options={options}
       />
+      <PromptComponent ref={modalRef} />
     </NoSsr>
   );
 }
